refactor: Spring WebFlux から MVC + Virtual Thread へ移行
WebFlux + Functional Routing による冗長さを解消し、JDK 21 の Virtual Thread を
活用した同期コード前提の構成に切り替える。

- build.gradle.kts: spring-boot-starter-webflux を spring-boot-starter-web に
  差し替え。Reactor / kotlinx-coroutines-reactor / WebTestClient / reactor-test
  / kotlinx-coroutines-test を削除。MVC スライステスト用に
  spring-boot-starter-webmvc-test、統合テスト用に spring-boot-resttestclient を追加
- gradle/libs.versions.toml: springdoc の starter を webflux 版から webmvc 版に変更
- src/main/kotlin/com/example/api/ApiApplication.kt: 削除する RouterConfig から
  @OpenAPIDefinition を移動
- src/main/kotlin/com/example/api/controller/HelloController.kt: @RestController
  + @GetMapping ベースの実装として新設
- src/main/kotlin/com/example/api/config/RouterConfig.kt: coRouter ベースの
  ルーティング定義は不要のため削除
- src/main/kotlin/com/example/api/handler/HelloHandler.kt: HelloController に
  置き換えたため削除
- src/main/resources/application.yml: spring.threads.virtual.enabled を true に
  設定し Tomcat のリクエスト処理スレッドを Virtual Thread 化
- src/test/kotlin/com/example/api/controller/HelloControllerTest.kt: @WebFluxTest
  + WebTestClient から @WebMvcTest + MockMvcTester に書き換え
- src/test/kotlin/com/example/api/OpenApiTest.kt: WebTestClient から
  RestTestClient (RestClient ベースの sync 版) に書き換え
- src/test/kotlin/com/example/api/actuator/HealthEndpointTest.kt: 同上の理由で
  RestTestClient に書き換え
