+++ src/main/kotlin/com/example/api/ApiApplication.kt
@@ -1,9 +1,18 @@
 package com.example.api
 
+import io.swagger.v3.oas.annotations.OpenAPIDefinition
+import io.swagger.v3.oas.annotations.info.Info
+import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
+@OpenAPIDefinition(
+    info = Info(title = "toy-box", summary = "toy-box の API 定義です。"),
+    tags = [
+        Tag(name = "Hello", description = "サンプル"),
+    ],
+)
 class ApiApplication
 
 fun main(args: Array<String>) {
