+++ src/test/kotlin/com/example/api/actuator/HealthEndpointTest.kt
@@ -1,21 +1,21 @@
 package com.example.api.actuator
 
 import org.junit.jupiter.api.Test
+import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureRestTestClient
 import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.boot.webtestclient.autoconfigure.AutoConfigureWebTestClient
 import org.springframework.test.context.TestConstructor
 import org.springframework.test.context.TestConstructor.AutowireMode
-import org.springframework.test.web.reactive.server.WebTestClient
+import org.springframework.test.web.servlet.client.RestTestClient
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureWebTestClient
+@AutoConfigureRestTestClient
 @TestConstructor(autowireMode = AutowireMode.ALL)
 class HealthEndpointTest(
-    val webTestClient: WebTestClient,
+    val restTestClient: RestTestClient,
 ) {
     @Test
     fun `ヘルスエンドポイントを呼び出すとUPステータスが返される`() {
-        webTestClient
+        restTestClient
             .get()
             .uri("/actuator/health")
             .exchange()
